Dues/requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.0
celery==5.4.0
cffi==1.17.0
channels==4.1.0
channels-redis==4.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==43.0.0
daphne==4.1.2
Django==5.1
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-redis==5.4.0
django-timezone-field==7.0
djangorestframework==3.15.2
hyperlink==21.0.0
idna==3.8
incremental==24.7.2
kombu==5.4.1
msgpack==1.0.8
passlib==1.7.4
pillow==10.4.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.8
service-identity==24.1.0
setuptools==73.0.1
six==1.16.0
sqlparse==0.5.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.0.1
